+++ config/config.exs
@@ -102,14 +102,9 @@ config :esbuild,
          --external:/fonts/*
          --external:/images/*
          js/app.js
-         js/editor/Editor.tsx
          js/react/components/DataclipViewer.tsx
          js/react/components/CollectionPreviewViewer.tsx
-         js/job-editor/JobEditor.tsx
-         js/workflow-editor/WorkflowEditor.tsx
-         js/workflow-store/WorkflowStore.tsx
          js/manual-run-panel/ManualRunPanel.tsx
-         js/panel/panels/WorkflowRunPanel.tsx
          js/collaborative-editor/CollaborativeEditor.tsx
          js/picker/Picker.tsx
          js/picker/PickerButton.tsx