+++ lib/lightning/config.ex
@@ -322,15 +322,6 @@ defmodule Lightning.Config do
     end
 
     @impl true
-    def ui_metrics_tracking_enabled? do
-      Keyword.get(ui_metrics_tracking_config(), :enabled)
-    end
-
-    defp ui_metrics_tracking_config do
-      Application.get_env(:lightning, :ui_metrics_tracking, [])
-    end
-
-    @impl true
     def credential_transfer_token_validity_in_days do
       2
     end
@@ -368,14 +359,6 @@ defmodule Lightning.Config do
     def external_metrics_module do
       Application.get_env(:lightning, Lightning.Extensions, [])
       |> Keyword.get(:external_metrics)
-    end
-
-    @impl true
-    def ai_assistant_modes do
-      %{
-        job: LightningWeb.Live.AiAssistant.Modes.JobCode,
-        workflow: LightningWeb.Live.AiAssistant.Modes.WorkflowTemplate
-      }
     end
 
     @impl true
@@ -529,7 +512,6 @@ defmodule Lightning.Config do
   @callback storage() :: term()
   @callback storage(key :: atom()) :: term()
   @callback token_signer() :: Joken.Signer.t()
-  @callback ui_metrics_tracking_enabled?() :: boolean()
   @callback usage_tracking() :: Keyword.t()
   @callback usage_tracking_cleartext_uuids_enabled?() :: boolean()
   @callback usage_tracking_cron_opts() :: [Oban.Plugins.Cron.cron_input()]
@@ -546,7 +528,6 @@ defmodule Lightning.Config do
   @callback gdpr_banner() :: map() | false
   @callback gdpr_preferences() :: map() | false
   @callback external_metrics_module() :: module() | nil
-  @callback ai_assistant_modes() :: %{atom() => module()}
   @callback per_workflow_claim_limit() :: pos_integer()
   @callback claim_work_mem() :: String.t() | nil
   @callback log_queue_queries() :: boolean()
@@ -786,10 +767,6 @@ defmodule Lightning.Config do
     impl().promex_enabled?()
   end
 
-  def ui_metrics_tracking_enabled? do
-    impl().ui_metrics_tracking_enabled?()
-  end
-
   def credential_transfer_token_validity_in_days do
     impl().credential_transfer_token_validity_in_days()
   end
@@ -816,10 +793,6 @@ defmodule Lightning.Config do
 
   def external_metrics_module do
     impl().external_metrics_module()
-  end
-
-  def ai_assistant_modes do
-    impl().ai_assistant_modes()
   end
 
   def metrics_run_performance_age_seconds do