+++ lib/lightning_web/channels/workflow_channel.ex
@@ -185,20 +185,6 @@ defmodule LightningWeb.WorkflowChannel do
   end
 
   @impl true
-  def handle_in("switch_to_legacy_editor", _payload, socket) do
-    user = socket.assigns[:current_user]
-
-    # Set switch to legacy to true
-    Lightning.Accounts.update_user_preference(
-      user,
-      "prefer_legacy_editor",
-      true
-    )
-
-    {:reply, {:ok, %{}}, socket}
-  end
-
-  @impl true
   def handle_in("get_context", _payload, socket) do
     user = socket.assigns[:current_user]
     workflow = socket.assigns.workflow